Use base template to dislay books so that shortcode works
