--- src/_includes/book.njk
@@ -1,23 +1,0 @@
-<li>
-  <details>
-    <summary>
-      {% set coverUri %}./src/covers/{{ book.fileSlug }}.jpg{% endset %}
-      {% responsiveImage coverUri, "" %}
-      <div{% if book.data.language != "en" %} lang="{{ book.data.language }}"{% endif %}>
-        <strong>{{ book.data.title | safe }}</strong> 
-        {% if book.data.language == "nl" %}door{% else %}by{% endif %} {{ book.data.author }}
-      </div>
-    </summary>
-    <p>
-      {{ book.data.rating }}<span aria-hidden="true">✯</span><span class="visually-hidden">stars</span> 
-      <span aria-hidden="true">・</span>
-      {{ book.data.pages }} pages 
-      <span aria-hidden="true">・</span> 
-      {{ book.data.publisher }} ({{ book.data.publishYear }})
-      {% if book.data.authorSite %}
-      <span aria-hidden="true">・</span> 
-      <a href="{{ book.data.authorSite }}">Author's site</a>
-      {% endif %}
-    </p>
-  </details>
-</li>
