--- src/_includes/year.njk
@@ -1,13 +1,0 @@
-<h2 id="{{ year }}"><span>{{ year }}</span></h2>
-<span hidden>({{ collections[year].length }})</span>
-<ul>
-{%- asyncEach book in collections[year] -%}
-{% if filterLanguage %}
-  {% if book.data.language == filterLanguage %}
-    {% include "book.njk" %}
-  {% endif %}
-{% else %}
-  {% include "book.njk" %}
-{% endif %}
-{%- endeach -%}
-</ul>
