+++ src/_routes/english.njk
@@ -1,10 +1,5 @@
 ---
 layout: base
 permalink: "in/english/"
-title: Hidde's list of books in English
+title: Hidde's list of books in Dutch
 ---
-
-{% for year in years %}
-{% set filterLanguage = "en" %}
-{% include "year.njk" %}
-{% endfor %}
