+++ src/_routes/nederlands.njk
@@ -3,8 +3,3 @@ layout: base
 permalink: "in/nederlands/"
 title: Hidde's list of books in Dutch
 ---
-
-{% for year in years %}
-{% set filterLanguage = "nl" %}
-{% include "year.njk" %}
-{% endfor %}
