+++ src/index.njk
@@ -5,10 +5,8 @@ pageClass: home
 ---
 
 <aside>
-  <a class="hiddedevriesnl" href="https://hiddedevries.nl"><img src="/favicon.png" alt="hiddedevries.nl"></a>
+  <a class="hiddedevriesnl" href="https://hiddedevries.nl/en"><img src="/favicon.png" alt="hiddedevries.nl"></a>
   <p><a href="#2020">2020</a> <span aria-hidden="true">・</span> <a href="#2019">2019</a> <span aria-hidden="true">・</span> <a href="#2018">2018</a> <span aria-hidden="true">・</span> <a href="#2017">2017</a> <span aria-hidden="true">・</span> <a href="https://github.com/hidde/books">Source code</a></p>
   <p>Copyright of book covers lies with their respective owners, displayed under fair use.</p>
 </aside>
-{% asyncEach year in years %}
-{% include "year.njk" %}
-{% endeach %}
+
